implemented childAlertURL
